Add files via upload
This is a simple Python command-line game that allows a user to play Rock, Paper, Scissors against the computer. The player chooses between rock, paper, or scissors, and the computer makes a random choice. The game then compares both choices to determine the winner based on the classic rules:

Rock beats Scissors

Scissors beats Paper

Paper beats Rock

The program uses the random module to generate the computer's choice and input() to get the user's choice. It then prints the result (win, lose, or draw) and optionally allows the user to play again.